Move all exported constants to a separate file
