+++ src/parser/jobs/blm/modules/RotationWatchdog.tsx
@@ -11,11 +11,9 @@ import {History, HistoryEntry} from 'parser/core/modules/ActionWindow/History'
 import {HistoryEntryPredicate} from 'parser/core/modules/ActionWindow/windows/ActionWindow'
 import {Actors} from 'parser/core/modules/Actors'
 import {Invulnerability} from 'parser/core/modules/Invulnerability'
-import {SEVERITY} from 'parser/core/modules/Suggestions'
 import {UnableToAct} from 'parser/core/modules/UnableToAct'
-import React, {Fragment, ReactNode} from 'react'
+import React, {Fragment} from 'react'
 import {Icon, Message} from 'semantic-ui-react'
-import {ensureRecord} from 'utilities'
 import DISPLAY_ORDER from './DISPLAY_ORDER'
 import {FIRE_SPELLS, ICE_SPELLS} from './Elements'
 import {ASTRAL_UMBRAL_DURATION, ASTRAL_UMBRAL_MAX_STACKS, BLMGaugeState, UMBRAL_HEARTS_MAX_STACKS} from './Gauge'
@@ -32,27 +30,16 @@ import {RotationErrorNotesEvaluator} from './RotationWatchdog/RotationErrorNotes
 import {SkipB4Evaluator} from './RotationWatchdog/SkipB4Evaluator'
 import {SkipT3Evaluator} from './RotationWatchdog/SkipT3Evaluator'
 import {UptimeSoulsEvaluator} from './RotationWatchdog/UptimeSoulsEvaluator'
+import {CycleMetadata, ROTATION_ERRORS, HIDDEN_PRIORITY_THRESHOLD} from './RotationWatchdog/WatchdogConstants'
 
 const DEBUG_SHOW_ALL = false && process.env.NODE_ENV !== 'production'
 
 const MAX_POSSIBLE_FIRE4 = 6
-export const NO_UH_EXPECTED_FIRE4 = 4
+const NO_UH_EXPECTED_FIRE4 = 4
 const MAX_MP = 10000
 const EXTRA_CASTS_FROM_MANAFONT = 1
 
 const EXTRA_F4_COP_THRESHOLD = 0.5 // Feelycraft
-
-export const DEFAULT_SEVERITY_TIERS = {
-	1: SEVERITY.MINOR,
-	3: SEVERITY.MEDIUM,
-	5: SEVERITY.MAJOR,
-}
-
-export const ENHANCED_SEVERITY_TIERS = {
-	1: SEVERITY.MINOR,
-	2: SEVERITY.MEDIUM,
-	3: SEVERITY.MAJOR,
-}
 
 const ROTATION_ENDPOINTS: ActionKey[] = [
 	'BLIZZARD_III',
@@ -62,59 +49,9 @@ const ROTATION_ENDPOINTS: ActionKey[] = [
 ]
 
 // This is feelycraft at the moment. Rotations shorter than this won't be processed for errors.
-export const MIN_ROTATION_LENGTH = 3
+const MIN_ROTATION_LENGTH = 3
 
-export interface CycleErrorCode {priority: number, message: ReactNode}
-export const DEATH_PRIORITY = 101 // Define this const here so we can reference it in both classes
-export const HIDDEN_PRIORITY_THRESHOLD = 2 // Same as ^
-/**
- * Error type codes, higher values indicate higher priority errors. If you add more, adjust the IDs to ensure correct priorities.
- * Only the highest priority error will be displayed in the 'Reason' column.
- * NOTE: Cycles with values at or below HIDDEN_PRIORITY_THRESHOLD will be filtered out of the RotationTable display
- * unless the DEBUG_SHOW_ALL variable is set to true
- */
-export const ROTATION_ERRORS = ensureRecord<CycleErrorCode>()({
-	NO_ERROR: {priority: 0, message: 'No errors'},
-	FINAL_OR_DOWNTIME: {priority: 1, message: 'Ended with downtime, or last cycle'},
-	SHORT: {priority: HIDDEN_PRIORITY_THRESHOLD, message: 'Too short, won\'t process'},
-	// Messages below should be Trans objects since they'll be displayed to end users
-	SHOULD_SKIP_T3: {priority: 8, message: <Trans id="blm.rotation-watchdog.error-messages.should-skip-t3">Should skip hardcast <DataLink action="THUNDER_III"/></Trans>},
-	SHOULD_SKIP_B4: {priority: 9, message: <Trans id="blm.rotation-watchdog.error-messages.should-skip-b4">Should skip <DataLink action="BLIZZARD_IV"/></Trans>},
-	MISSING_FIRE4S: {priority: 10, message: <Trans id="blm.rotation-watchdog.error-messages.missing-fire4s">Missing one or more <DataLink action="FIRE_IV"/>s</Trans>}, // These two errors are lower priority since they can be determined by looking at the
-	MISSED_ICE_PARADOX: {priority: 15, message: <Trans id="blm.rotation-watchdog.error-messages.missed-ice-paradox">Missed <DataLink action="PARADOX"/> in Umbral Ice</Trans>},
-	MISSING_DESPAIRS: {priority: 20, message: <Trans id="blm.rotation-watchdog.error-messages.missing-despair">Missing one or more <DataLink action="DESPAIR"/>s</Trans>}, // target columns in the table, so we want to tell players about other errors first
-	MANAFONT_BEFORE_DESPAIR: {priority: 40, message: <Trans id="blm.rotation-watchdog.error-messages.manafont-before-despair"><DataLink action="MANAFONT"/> used before <DataLink action="DESPAIR"/></Trans>},
-	EXTRA_T3: {priority: 59, message: <Trans id="blm.rotation-watchdog.error-messages.extra-t3">Extra <DataLink action="THUNDER_III"/>s</Trans>}, // Extra T3 and Extra F1 are *very* similar in terms of per-GCD potency loss
-	EXTRA_F1: {priority: 60, message: <Trans id="blm.rotation-watchdog.error-messages.extra-f1">Extra <DataLink action="FIRE_I"/></Trans>}, // These two codes should stay close to each other
-	NO_FIRE_SPELLS: {priority: 80, message: <Trans id="blm.rotation-watchdog.error-messages.no-fire-spells">Rotation included no Fire spells</Trans>},
-	DROPPED_AF_UI: {priority: 100, message: <Trans id="blm.rotation-watchdog.error-messages.dropped-astral-umbral">Dropped Astral Fire or Umbral Ice</Trans>},
-	DIED: {priority: DEATH_PRIORITY, message: <Trans id="blm.rotation-watchdog.error-messages.died"><DataLink showName={false} action="RAISE"/> Died</Trans>},
-})
-
-export interface CycleMetadata {
-	errorCode: CycleErrorCode
-	finalOrDowntime: boolean
-	missingDespairs: boolean
-	missingFire4s: boolean
-	wasTPF1: boolean
-	expectedFire4sBeforeDespair: number
-	expectedFire4s: number,
-	expectedDespairs: number,
-	hardT3sInFireCount: number
-	firePhaseMetadata: PhaseMetadata
-}
-
-export interface PhaseMetadata {
-	startTime: number
-	initialMP: number
-	initialGaugeState: BLMGaugeState
-	fullElementTime: number
-	fullElementMP: number
-	fullElementGaugeState: BLMGaugeState
-	circleOfPowerPct: number
-}
-
-export const EMPTY_GAUGE_STATE: BLMGaugeState = {
+const EMPTY_GAUGE_STATE: BLMGaugeState = {
 	astralFire: 0,
 	umbralIce: 0,
 	umbralHearts: 0,
@@ -265,6 +202,7 @@ export class RotationWatchdog extends RestartWindow {
 		this.addEvaluator(new SkipB4Evaluator({
 			blizzard4Id: this.data.actions.BLIZZARD_IV.id,
 			fire4action: this.data.actions.FIRE_IV,
+			minExpectedF4s: NO_UH_EXPECTED_FIRE4,
 			metadataHistory: this.metadataHistory,
 		}))
 		//#endregion
