+++ src/parser/jobs/blm/modules/RotationWatchdog/ExtraF1Evaluator.tsx
@@ -3,9 +3,10 @@ import {DataLink} from 'components/ui/DbLink'
 import {EvaluatedAction, WindowEvaluator} from 'parser/core/modules/ActionWindow'
 import {History, HistoryEntry} from 'parser/core/modules/ActionWindow/History'
 import {TieredSuggestion} from 'parser/core/modules/Suggestions'
+import {DEFAULT_SEVERITY_TIERS} from 'parser/jobs/dnc/CommonData'
 import React from 'react'
-import {ROTATION_ERRORS, DEFAULT_SEVERITY_TIERS, CycleMetadata} from '../RotationWatchdog'
 import {assignErrorCode, getMetadataForWindow} from './EvaluatorUtilities'
+import {CycleMetadata, ROTATION_ERRORS} from './WatchdogConstants'
 
 export interface ExtraF1EvaluatorOpts {
 	suggestionIcon: string
