+++ src/parser/jobs/blm/modules/RotationWatchdog/ExtraHardT3Evaluator.tsx
@@ -4,10 +4,11 @@ import {EvaluatedAction, WindowEvaluator} from 'parser/core/modules/ActionWindow
 import {History, HistoryEntry} from 'parser/core/modules/ActionWindow/History'
 import {Data} from 'parser/core/modules/Data'
 import {TieredSuggestion} from 'parser/core/modules/Suggestions'
+import {DEFAULT_SEVERITY_TIERS} from 'parser/jobs/dnc/CommonData'
 import React from 'react'
 import Procs from '../Procs'
-import {ROTATION_ERRORS, DEFAULT_SEVERITY_TIERS, CycleMetadata} from '../RotationWatchdog'
 import {assignErrorCode, getMetadataForWindow} from './EvaluatorUtilities'
+import {CycleMetadata, ROTATION_ERRORS} from './WatchdogConstants'
 
 export interface ExtraHardT3EvaluatorOpts {
 	data: Data,
