+++ src/parser/jobs/blm/modules/RotationWatchdog/ManafontTimingEvaluator.tsx
@@ -6,8 +6,8 @@ import {RulePassedEvaluator} from 'parser/core/modules/ActionWindow/evaluators/R
 import {History, HistoryEntry} from 'parser/core/modules/ActionWindow/History'
 import {TieredSuggestion} from 'parser/core/modules/Suggestions'
 import React from 'react'
-import {ROTATION_ERRORS, ENHANCED_SEVERITY_TIERS, CycleMetadata} from '../RotationWatchdog'
 import {assignErrorCode} from './EvaluatorUtilities'
+import {CycleMetadata, ROTATION_ERRORS, ENHANCED_SEVERITY_TIERS} from './WatchdogConstants'
 
 export interface ManafontTimingEvaluatorOpts {
 	manafontAction: Action
