+++ src/parser/jobs/blm/modules/RotationWatchdog/RotationErrorNotesEvaluator.tsx
@@ -2,8 +2,8 @@ import {Trans} from '@lingui/macro'
 import {EvaluatedAction, NotesEvaluator} from 'parser/core/modules/ActionWindow'
 import {History, HistoryEntry} from 'parser/core/modules/ActionWindow/History'
 import React from 'react'
-import {CycleMetadata} from '../RotationWatchdog'
 import {getMetadataForWindow} from './EvaluatorUtilities'
+import {CycleMetadata} from './WatchdogConstants'
 
 export class RotationErrorNotesEvaluator extends NotesEvaluator {
 	private metadataHistory: History<CycleMetadata>
