+++ src/parser/jobs/blm/modules/RotationWatchdog/SkipT3Evaluator.tsx
@@ -4,8 +4,8 @@ import {EvaluatedAction, WindowEvaluator} from 'parser/core/modules/ActionWindow
 import {History, HistoryEntry} from 'parser/core/modules/ActionWindow/History'
 import {TieredSuggestion} from 'parser/core/modules/Suggestions'
 import React from 'react'
-import {ROTATION_ERRORS, ENHANCED_SEVERITY_TIERS, CycleMetadata} from '../RotationWatchdog'
 import {assignErrorCode, getMetadataForWindow} from './EvaluatorUtilities'
+import {CycleMetadata, ROTATION_ERRORS, ENHANCED_SEVERITY_TIERS} from './WatchdogConstants'
 
 export interface SkipT3EvaluatorOpts {
 	suggestionIcon: string
