+++ src/parser/jobs/blm/modules/RotationWatchdog/UptimeSoulsEvaluator.tsx
@@ -6,7 +6,7 @@ import {HistoryEntry} from 'parser/core/modules/ActionWindow/History'
 import {Invulnerability} from 'parser/core/modules/Invulnerability'
 import {TieredSuggestion} from 'parser/core/modules/Suggestions'
 import React from 'react'
-import {DEFAULT_SEVERITY_TIERS} from '../RotationWatchdog'
+import {DEFAULT_SEVERITY_TIERS} from './WatchdogConstants'
 
 export interface UptimeSoulsEvaluatorOpts {
 	umbralSoulAction: Action
